// app/src/generated/models/SOPsModel.ts
/*!
 * Copyright (C) Microsoft Corporation. All rights reserved.
 * This file is autogenerated. Do not edit this file directly.
 */

export interface EditorValue {
  "@odata.type": string;
  Claims: string;
  DisplayName: string;
  Email: string;
  Picture: string;
  Department: string;
  JobTitle: string;
}

export interface SOPStatusValue {
  "@odata.type": string;
  Value: string;
  Id: number;
}

export interface AuthorValue {
  "@odata.type": string;
  Claims: string;
  DisplayName: string;
  Email: string;
  Picture: string;
  Department: string;
  JobTitle: string;
}

export interface CheckoutUserValue {
  "@odata.type": string;
  Claims: string;
  DisplayName: string;
  Email: string;
  Picture: string;
  Department: string;
  JobTitle: string;
}

export interface SOPsBase {
  ID?: number;
  Title?: string;
  Modified?: string;
  "Editor#Claims"?: string;
  Department?: string;
  ProcessOwner?: string;
  "SOPStatus#Id"?: number;
  SOPVersion?: string;
  PromptVersion?: string;
  OwnerComments?: string;
  CurrentBatch?: number;
  OData__ColorTag?: string;
  ComplianceAssetId?: string;
  OData__ExtendedDescription?: string;
  Created?: string;
  "Author#Claims"?: string;
  "CheckoutUser#Claims"?: string;
  OData__DisplayName?: string;
  "{Identifier}"?: string;
  "{IsFolder}"?: boolean;
  "{Thumbnail}"?: {
    "@odata.type"?: string;
    // URL to the small, highly compressed size thumbnail of the item, if available.
    Small?: string;
    // URL to the standard size thumbnail of the item, if available.
    Medium?: string;
    // URL to the largest size thumbnail of the item, if available.
    Large?: string;
  };
  "{Link}"?: string;
  "{Name}"?: string;
  "{FilenameWithExtension}"?: string;
  "{Path}"?: string;
  "{FullPath}"?: string;
  "{ModerationStatus}"?: string;
  "{ModerationComment}"?: string;
  "{ContentType}#Id"?: string;
  "{ContentType}"?: {
    "@odata.type"?: string;
    Id?: string;
    Name?: string;
  };
  "{IsCheckedOut}"?: boolean;
  "{DriveId}"?: string;
  "{DriveItemId}"?: string;
  "{SensitivityLabelId}"?: string;
  "{SensitivityLabelName}"?: string;
  "{VersionNumber}"?: string;
  "{TriggerWindowStartToken}"?: string;
  "{TriggerWindowEndToken}"?: string;
}

export interface SOPsRead extends SOPsBase {
  Editor?: EditorValue;
  SOPStatus?: SOPStatusValue;
  Author?: AuthorValue;
  CheckoutUser?: CheckoutUserValue;
}

export interface SOPsWrite extends SOPsBase {
  Editor?: EditorValue;
  SOPStatus?: string;
  Author?: AuthorValue;
  CheckoutUser?: CheckoutUserValue;
}